chore(switch): toggler component done
