--- lib/components/DefaultToggleSwitch/index.module.css
@@ -1,7 +1,0 @@
-.DefaultToggleSwitch {
-  display: inline-flex;
-  position: relative;
-  align-items: center;
-  cursor: pointer;
-}
-
