+++ lib/components/DefaultToggleSwitch/index.tsx
@@ -1,82 +1,108 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/ban-ts-comment */
-import clsx from 'clsx';
-import React, { useState } from 'react';
+import { type ClassValue, clsx } from 'clsx';
+import React from 'react';
 import { twMerge } from 'tailwind-merge';
-import styles from './index.module.css';
-const variants = {
-  primary: styles.DefaultToggleSwitch,
-  small: 'h-7 w-14 min-w-14',
-  medium: 'h-14 w-28 min-w-28',
-  large: 'h-28 w-56 min-w-56',
-  green: 'bg-emerald-100 text-emerald-800 border-emerald-800',
-  yellow: 'bg-yellow-100 text-yellow-800 border-yellow-800',
-  red: 'bg-red-100 text-red-800 border-red-800',
-  blue: 'bg-blue-100 text-blue-800 border-blue-800',
-} as const;
 
-//@ts-ignore
-export declare interface IDefaultToggleSwitchProps extends React.HTMLAttributes<never> {
-  base?: 'primary';
-  size?: 'small' | 'medium' | 'large';
-  color?: 'green' | 'yellow' | 'red' | 'blue';
-  initialChecked?: boolean;
-  label?: string;
-  handleToggle: any;
-  onChange: (checked: boolean) => void;
+function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
 }
+
+const switchColors = {
+  background: {
+    default: 'bg-neutral-200 dark:bg-neutral-700',
+    hover: 'hover:bg-neutral-300 dark:hover:bg-neutral-600',
+    checked:
+      'has-[input[type=checkbox]:checked]:bg-neutral-700 dark:has-[input[type=checkbox]:checked]:bg-neutral-300',
+    checkedHover:
+      'has-[input[type=checkbox]:checked]:hover:bg-black dark:has-[input[type=checkbox]:checked]:hover:bg-white',
+    disabled:
+      'has-[input[type=checkbox]:disabled]:bg-neutral-100 dark:has-[input[type=checkbox]:disabled]:bg-neutral-800',
+    disabledHover:
+      'has-[input[type=checkbox]:disabled]:hover:bg-neutral-100 dark:has-[input[type=checkbox]:disabled]:hover:bg-neutral-800',
+    disabledChecked:
+      'has-[input[type=checkbox]:disabled:checked]:bg-neutral-300 dark:has-[input[type=checkbox]:disabled:checked]:bg-neutral-700',
+    disabledCheckedHover:
+      'has-[input[type=checkbox]:disabled:checked]:hover:bg-neutral-300 dark:has-[input[type=checkbox]:disabled:checked]:hover:bg-neutral-700',
+  },
+  button: {
+    default: 'bg-white dark:bg-black',
+  },
+};
+const switchSizes = {
+  wrapper: {
+    default:
+      'w-10 min-w-10 has-[input[type=checkbox]:not(:checked)]:pr-5 has-[input[type=checkbox]:checked]:pl-5',
+    small:
+      'h-10 w-18 min-w-18 has-[input[type=checkbox]:not(:checked)]:pr-12 has-[input[type=checkbox]:checked]:pl-12',
+    medium:
+      'h-14 w-26 min-w-26 has-[input[type=checkbox]:not(:checked)]:pr-14 has-[input[type=checkbox]:checked]:pl-14',
+    large:
+      'h-26 w-58 min-w-58 has-[input[type=checkbox]:not(:checked)]:pr-32 has-[input[type=checkbox]:checked]:pl-32',
+  },
+  button: {
+    default: 'w-4 h-4 rounded-full',
+    small: 'h-8 w-8 rounded-full',
+    medium: 'h-12 w-12 rounded-full',
+    large: 'h-24 w-24 rounded-full',
+  },
+};
+export interface IDefaultToggleSwitchProps extends React.AllHTMLAttributes<HTMLInputElement> {}
+
 export const DefaultToggleSwitch = React.forwardRef<HTMLInputElement, IDefaultToggleSwitchProps>(
   (
     {
-      initialChecked = false,
-      className = styles.DefaultToggleSwitch,
-      color = 'success',
-      label,
-      size = 'small',
+      className,
+      size = 'default',
+      type = 'checkbox',
+      defaultChecked,
       onChange,
-      base = 'primary',
+      checked: propChecked,
+      ...props
     },
     ref,
   ) => {
-    const [checked, setChecked] = useState(initialChecked);
-    const handleChange = (): void => {
-      const newChecked = !checked;
-      setChecked(newChecked);
-      if (onChange) onChange(newChecked);
-    };
-    const toggle = (): void => onChange(!checked);
-    const handleClick = (): void => {
-      toggle();
-    };
-    const labelStyle = `ml-3 text-sm font-medium text-[#0B2447] dark:text-white`;
-    const switchOuterStyle = `w-14 h-7 bg-white dark:bg-gray-700 border-2 border-[#0B2447] dark:border-white peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-[#0B2447]/50 dark:peer-focus:ring-white/50 rounded-full peer transition-all duration-300 ${checked ? 'shadow-[3px_3px_0px_0px_#0B2447] dark:shadow-[3px_3px_0px_0px_#ffffff]' : ''}`;
-    const switchInnerStyle = `absolute top-1 left-1 bg-[#0B2447] dark:bg-white rounded-full ${size !== 'small' ? (size !== 'medium' ? 'h-24 w-24' : 'h-12 w-12') : 'h-5 w-5'} transition-all duration-300 ${checked ? (size !== 'small' ? (size !== 'medium' ? 'translate-x-28' : 'translate-x-14') : 'translate-x-7 ') : ''}`;
+    const [checked, setChecked] = React.useState(defaultChecked ?? false);
+    React.useEffect(() => {
+      if (typeof propChecked === 'boolean') {
+        setChecked(propChecked);
+      }
+    }, [propChecked]);
+    const internalRef = React.useRef<HTMLInputElement | null>(null);
+    const renderWrapperStyles = `${size === 'default' ? switchSizes.wrapper.default : size !== 'small' ? (size !== 'medium' ? switchSizes.wrapper.large : switchSizes.wrapper.medium) : switchSizes.wrapper.small} relative inline-block group/switch rounded-full p-1 ${switchColors.background.default} ${switchColors.background.hover} ${switchColors.background.checked} ${switchColors.background.checkedHover} ${switchColors.background.disabled} ${switchColors.background.disabledHover} ${switchColors.background.disabledChecked} ${switchColors.background.disabledCheckedHover} has-[input[type=checkbox]:disabled]:cursor-not-allowed transition-all duration-200 ease-in-out`;
+    const renderButtonStyles = `${size === 'default' ? switchSizes.button.default : size !== 'small' ? (size !== 'medium' ? switchSizes.button.large : switchSizes.button.medium) : switchSizes.button.small}`;
+
     return (
-      <>
-        <label className={twMerge(clsx(variants[base as keyof typeof variants]))}>
-          <input
-            type="checkbox"
-            ref={ref}
-            className="sr-only peer"
-            checked={checked}
-            onClick={handleClick}
-            onChange={handleChange}
-          />
-          <div
-            className={twMerge(
-              switchOuterStyle,
-              clsx(
-                variants[size as keyof typeof variants],
-                variants[color as keyof typeof variants],
-              ),
-              className,
-            )}
-          >
-            <div className={switchInnerStyle}></div>
-          </div>
-          <span className={labelStyle}>{label}</span>
-        </label>
-      </>
+      <label
+        htmlFor={`defaultToggleSwitchId`}
+        className={cn(renderWrapperStyles, className)}
+      >
+        <input
+          id={`defaultToggleSwitchId`}
+          defaultChecked={defaultChecked}
+          checked={typeof defaultChecked === 'boolean' ? undefined : checked}
+          type={type}
+          className="hidden"
+          onChange={(e) => {
+            setChecked(e.currentTarget.checked);
+            onChange?.(e);
+          }}
+          ref={(el) => {
+            internalRef.current = el;
+            if (typeof ref === 'function') {
+              ref(el);
+            } else if (ref) {
+              ref.current = el;
+            }
+          }}
+          {...props}
+        />
+        <div
+          className={cn(
+            `${switchSizes.button.default}`,
+            `${switchColors.button.default}`,
+            `${renderButtonStyles}`,
+          )}
+        />
+      </label>
     );
   },
 );
