--- lib/components/StaticToggleButton/index.stories.tsx
@@ -1,12 +1,0 @@
-import { StoryFn } from '@storybook/react';
-import { StaticToggleButton } from './index';
-
-export default {
-  component: StaticToggleButton,
-};
-
-export const Default: StoryFn = (args) => <StaticToggleButton {...args} />;
-Default.args = {
-  type: 'checkbox',
-  defaultChecked: true,
-};
