+++ lib/components/UncontrolledToggleSwitch/index.stories.tsx
@@ -5,19 +5,20 @@ export default {
   component: UncontrolledToggleSwitch,
 };
 
-export const Toggled: StoryFn<IUncontrolledToggleSwitchProps> = (args) => (
-  <UncontrolledToggleSwitch {...args} />
-);
-Toggled.args = {
-  label: 'UncontrolledToggleSwitch label',
-  checked: true,
-  onChange: () => {},
-};
 export const NotToggled: StoryFn<IUncontrolledToggleSwitchProps> = (args) => (
   <UncontrolledToggleSwitch {...args} />
 );
 NotToggled.args = {
   label: 'UncontrolledToggleSwitch label',
   checked: false,
+  onChange: () => {},
+};
+
+export const Toggled: StoryFn<IUncontrolledToggleSwitchProps> = (args) => (
+  <UncontrolledToggleSwitch {...args} />
+);
+Toggled.args = {
+  label: 'UncontrolledToggleSwitch label',
+  checked: true,
   onChange: () => {},
 };
