+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@314oner_npm/universal-components-library",
-  "version": "3.4.4",
+  "version": "3.4.5",
   "type": "module",
   "main": "dist/main.js",
   "types": "dist/main.d.ts",
